Adds test for inspect_data_process_definition
